rename operator.py and test_operator.py to avoid conflicts with the Python standard library
