+++ tests/quara/objects/test_operators.py
@@ -6,15 +6,11 @@ from quara.objects import matrix_basis
 from quara.objects.composite_system import CompositeSystem
 from quara.objects.elemental_system import ElementalSystem
 from quara.objects.gate import Gate, get_h, get_i, get_s, get_x, get_y, get_z
-from quara.objects.operator import (
-    _composite,
-    _tensor_product,
-    _to_list,
-    composite,
-    tensor_product,
-)
+from quara.objects.operators import (_composite, _tensor_product, _to_list,
+                                     composite, tensor_product)
 from quara.objects.povm import Povm
-from quara.objects.state import State, get_x0_1q, get_x1_1q, get_z0_1q, get_z1_1q
+from quara.objects.state import (State, get_x0_1q, get_x1_1q, get_z0_1q,
+                                 get_z1_1q)
 
 
 def test_tensor_product_Gate_Gate():
